# docs/api/v2/V2_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API explorer for the go-deploy API. You can use it as a reference for the API endpoints.",
        "title": "go-deploy API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Support",
            "url": "https://github.com/kthcloud/go-deploy"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/kthcloud/go-deploy?tab=MIT-1-ov-file#readme"
        },
        "version": "1.0"
    },
    "paths": {
        "/v2/deployments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List deployments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deployment"
                ],
                "summary": "List deployments",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "List all",
                        "name": "all",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by user ID",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include shared",
                        "name": "shared",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/body.DeploymentRead"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create deployment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deployment"
                ],
                "summary": "Create deployment",
                "parameters": [
                    {
                        "description": "Deployment body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/body.DeploymentCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/body.DeploymentRead"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v2/deployments/{deploymentId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get deployment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deployment"
                ],
                "summary": "Get deployment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deployment ID",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/body.DeploymentRead"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update deployment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deployment"
                ],
                "summary": "Update deployment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deployment ID",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Deployment update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/body.DeploymentUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/body.DeploymentUpdated"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete deployment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deployment"
                ],
                "summary": "Delete deployment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deployment ID",
                        "name": "deploymentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/body.DeploymentCreated"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v2/hooks/harbor": {
            "post": {
                "description": "Handle Harbor hook",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deployment"
                ],
                "summary": "Handle Harbor hook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Basic auth token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "Harbor webhook body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/body.HarborWebhook"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "body.CustomDomainRead": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "body.DeploymentCreate": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "args": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 0,
                    "items": {
                        "type": "string"
                    }
                },
                "cpuCores": {
                    "type": "number",
                    "minimum": 0.1
                },
                "customDomain": {
                    "description": "CustomDomain is the domain that the deployment will be available on.\nThe max length is set to 243 to allow for a subdomain when confirming the domain.",
                    "type": "string"
                },
                "envs": {
                    "type": "array",
                    "maxItems": 1000,
                    "minItems": 0,
                    "items": {
                        "$ref": "#/definitions/body.Env"
                    }
                },
                "healthCheckPath": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 0
                },
                "image": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 1
                },
                "initCommands": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 0,
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 3
                },
                "private": {
                    "type": "boolean"
                },
                "ram": {
                    "type": "number",
                    "minimum": 0.1
                },
                "replicas": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0
                },
                "volumes": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 0,
                    "items": {
                        "$ref": "#/definitions/body.Volume"
                    }
                },
                "zone": {
                    "description": "Zone is the zone that the deployment will be created in.\nIf the zone is not set, the deployment will be created in the default zone.",
                    "type": "string"
                }
            }
        },
        "body.DeploymentCreated": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "jobId": {
                    "type": "string"
                }
            }
        },
        "body.DeploymentRead": {
            "type": "object",
            "properties": {
                "accessedAt": {
                    "type": "string"
                },
                "args": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "customDomain": {
                    "$ref": "#/definitions/body.CustomDomainRead"
                },
                "envs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/body.Env"
                    }
                },
                "error": {
                    "type": "string"
                },
                "healthCheckPath": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "initCommands": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "integrations": {
                    "description": "Integrations are currently not used, but could be used if we wanted to add a list of integrations to the deployment\n\nFor example GitHub",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "internalPort": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "ownerId": {
                    "type": "string"
                },
                "pingResult": {
                    "type": "integer"
                },
                "private": {
                    "type": "boolean"
                },
                "repairedAt": {
                    "type": "string"
                },
                "replicaStatus": {
                    "$ref": "#/definitions/body.ReplicaStatus"
                },
                "restartedAt": {
                    "type": "string"
                },
                "specs": {
                    "$ref": "#/definitions/body.DeploymentSpecs"
                },
                "status": {
                    "type": "string"
                },
                "storageUrl": {
                    "type": "string"
                },
                "teams": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "volumes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/body.Volume"
                    }
                },
                "zone": {
                    "type": "string"
                }
            }
        },
        "body.DeploymentSpecs": {
            "type": "object",
            "properties": {
                "cpuCores": {
                    "type": "number"
                },
                "ram": {
                    "type": "number"
                },
                "replicas": {
                    "type": "integer"
                }
            }
        },
        "body.DeploymentUpdate": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "args": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 0,
                    "items": {
                        "type": "string"
                    }
                },
                "cpuCores": {
                    "type": "number",
                    "minimum": 0.1
                },
                "customDomain": {
                    "description": "CustomDomain is the domain that the deployment will be available on.\nThe max length is set to 243 to allow for a subdomain when confirming the domain.",
                    "type": "string"
                },
                "envs": {
                    "type": "array",
                    "maxItems": 1000,
                    "minItems": 0,
                    "items": {
                        "$ref": "#/definitions/body.Env"
                    }
                },
                "healthCheckPath": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 0
                },
                "image": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 1
                },
                "initCommands": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 0,
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 3
                },
                "private": {
                    "type": "boolean"
                },
                "ram": {
                    "type": "number",
                    "minimum": 0.1
                },
                "replicas": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0
                },
                "volumes": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 0,
                    "items": {
                        "$ref": "#/definitions/body.Volume"
                    }
                }
            }
        },
        "body.DeploymentUpdated": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "jobId": {
                    "type": "string"
                }
            }
        },
...
        "body.Env": {
            "type": "object",
            "required": [
                "name",
                "value"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "value": {
                    "type": "string",
                    "maxLength": 10000,
                    "minLength": 1
                }
            }
        },
...
        "body.HarborWebhook": {
            "type": "object",
            "properties": {
                "event_data": {
                    "type": "object",
                    "properties": {
                        "repository": {
                            "type": "object",
                            "properties": {
                                "date_created": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "namespace": {
                                    "type": "string"
                                },
                                "repo_full_name": {
                                    "type": "string"
                                },
                                "repo_type": {
                                    "type": "string"
                                }
                            }
                        },
                        "resources": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "digest": {
                                        "type": "string"
                                    },
                                    "resource_url": {
                                        "type": "string"
                                    },
                                    "tag": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "occur_at": {
                    "type": "integer"
                },
                "operator": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "body.ReplicaStatus": {
            "type": "object",
            "properties": {
                "availableReplicas": {
                    "description": "AvailableReplicas is the number of replicas that are available.",
                    "type": "integer"
                },
                "desiredReplicas": {
                    "description": "DesiredReplicas is the number of replicas that the deployment should have.",
                    "type": "integer"
                },
                "readyReplicas": {
                    "description": "ReadyReplicas is the number of replicas that are ready.",
                    "type": "integer"
                },
                "unavailableReplicas": {
                    "description": "UnavailableReplicas is the number of replicas that are unavailable.",
                    "type": "integer"
                }
            }
        },
...
        "body.Volume": {
            "type": "object",
            "required": [
                "appPath",
                "name",
                "serverPath"
            ],
            "properties": {
                "appPath": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 3
                },
                "serverPath": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
        "sys.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "sys.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sys.Error"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-Api-Key",
            "in": "header"
        }
    }
}